refactor: utils의 constants를 constants 폴더로 이동 (#37)
