+++ src/components/modal/plan/PlanTag.tsx
@@ -8,8 +8,8 @@ import Input from '@/components/common/Input';
 import { TagIcon } from '@/components/icons';
 import ClassifierTitle from '@/components/sidebar/classifier/ClassifierTitle';
 import { MAX_CANDIDATE_LENGTH, MAX_TAG_LENGTH } from '@/constants';
+import { TAG_MOCK } from '@/constants/mock';
 import useDebounce from '@/hooks/useDebounce';
-import { TAG_MOCK } from '@/utils/mock';
 import { Candidate } from 'components/modal/plan/Candidates';
 
 const PlanTag: React.FC = () => {
