+++ src/hooks/rq/category.ts
@@ -1,8 +1,8 @@
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
+import { CATEGORY_MOCK } from '@/constants/mock';
+import { CATEGORY_KEY } from '@/constants/rqKeys';
 import { ICategory, ICategoryWithoutId } from '@/types/rq/category';
-import { CATEGORY_MOCK } from '@/utils/mock';
-import { CATEGORY_KEY } from '@/utils/rqKeys';
 
 // 카테고리 캐싱을 위한 키
 const KEY = [CATEGORY_KEY];
